clean up endpoint
